+++ src/freenet/clients/http/SimpleToadletServer.java
@@ -94,6 +94,7 @@ public final class SimpleToadletServer implements ToadletContainer, Runnable, Li
 	private boolean sendAllThemes;
 	private boolean advancedModeEnabled;
 	private final PageMaker pageMaker;
+	private boolean fetchKeyBoxAboveBookmarks;
 	
 	// Control
 	private Thread myThread;
@@ -259,6 +260,7 @@ public final class SimpleToadletServer implements ToadletContainer, Runnable, Li
 			NodeClientCore core = SimpleToadletServer.this.core;
 			if (core.node.pluginManager != null)
 				core.node.pluginManager.setFProxyTheme(cssTheme);
+			fetchKeyBoxAboveBookmarks = cssTheme.fetchKeyBoxAboveBookmarks;
 		}
 
 		@Override
@@ -485,6 +487,23 @@ public final class SimpleToadletServer implements ToadletContainer, Runnable, Li
 		
 		fproxyConfig.register("advancedModeEnabled", false, configItemOrder++, true, false, "SimpleToadletServer.advancedMode", "SimpleToadletServer.advancedModeLong",
 				new FProxyAdvancedModeEnabledCallback(this));
+
+		fproxyConfig.register("fetchKeyBoxAboveBookmarks", false, configItemOrder++,
+				true, false, "SimpleToadletServer.fetchKeyBoxAboveBookmarks",
+				"SimpleToadletServer.fetchKeyBoxAboveBookmarksLong", new BooleanCallback() {
+			@Override
+			public Boolean get() {
+				return fetchKeyBoxAboveBookmarks;
+			}
+
+			@Override
+			public void set(Boolean val) {
+				if(get().equals(val)) return;
+				fetchKeyBoxAboveBookmarks = val;
+			}
+		});
+		fetchKeyBoxAboveBookmarks = fproxyConfig.getBoolean("fetchKeyBoxAboveBookmarks");
+
 		fproxyConfig.register("enableExtendedMethodHandling", false, configItemOrder++, true, false, "SimpleToadletServer.enableExtendedMethodHandling", "SimpleToadletServer.enableExtendedMethodHandlingLong",
 				new BooleanCallback() {
 					@Override